Started on Backend API key gathering in .env file
