+++ .gitignore
@@ -1,0 +1,2 @@
+# Ignore API Key
+.env
