+++ src/alphavantage/Config.java
@@ -1,0 +1,14 @@
+package src.alphavantage;
+
+public class Config {
+    public static final String BASE_URL = "https://www.alphavantage.co/query?";
+    private String apiKey;
+
+    public Config(String apiKey) {
+        this.apiKey = apiKey;
+    }
+
+    public String getApiKey() {
+        return this.apiKey;
+    }
+}
